# docs/api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BASE Wealth Management API",
    "description": "REST API construida con Kotlin y Spring Boot 3 para gestión de activos patrimoniales y proyecciones financieras.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/api/v1/wealth/snapshots": {
      "get": {
        "tags": [
          "Wealth"
        ],
        "summary": "Serie histórica de patrimonio neto, con variación porcentual contra el punto anterior",
        "operationId": "getSnapshots",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SnapshotResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Wealth"
        ],
        "summary": "Captura un snapshot del patrimonio actual, calculado en el servidor",
        "operationId": "createSnapshot",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotResponse"
                }
              }
            }
          },
          "409": {
            "description": "Ya existe un snapshot para este segundo",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/wealth/estimate": {
      "get": {
        "tags": [
          "Wealth"
        ],
        "summary": "Calcula la proyección de interés compuesto y el progreso de hitos financieros",
        "operationId": "getEstimate",
        "parameters": [
          {
            "name": "contribution",
            "in": "query",
            "required": true,
            "schema": {
              "maximum": 1000000000,
              "exclusiveMaximum": false,
              "minimum": 0.0,
              "exclusiveMinimum": false,
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "yieldPct",
            "in": "query",
            "required": true,
            "schema": {
              "maximum": 100.0,
              "exclusiveMaximum": false,
              "minimum": 0.0,
              "exclusiveMinimum": false,
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "years",
            "in": "query",
            "required": true,
            "schema": {
              "maximum": 50,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "milestones",
            "in": "query",
            "required": false,
            "schema": {
              "maxItems": 5,
              "minItems": 0,
              "type": "array",
              "items": {
                "type": "number",
                "format": "double"
              }
            }
          },
          {
            "name": "principal",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "format": "double"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectionResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "SnapshotResponse": {
        "required": [
          "capturedAt",
          "id",
          "totalValueUsd"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "capturedAt": {
            "type": "string",
            "format": "date-time"
          },
          "totalValueUsd": {
            "type": "number",
            "format": "double"
          },
          "changePctFromPrevious": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "MilestoneResponse": {
        "required": [
          "amountUsd",
          "status"
        ],
        "type": "object",
        "properties": {
          "amountUsd": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "type": "string"
          },
          "monthsRequired": {
            "type": "integer",
            "format": "int32"
          },
          "targetMonth": {
            "type": "string"
          }
        }
      },
      "ProjectionPointResponse": {
        "required": [
          "futureValueUsd",
          "interestEarnedUsd",
          "totalContributedUsd",
          "year"
        ],
        "type": "object",
        "properties": {
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "futureValueUsd": {
            "type": "number",
            "format": "double"
          },
          "totalContributedUsd": {
            "type": "number",
            "format": "double"
          },
          "interestEarnedUsd": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "ProjectionResponse": {
        "required": [
          "annualYieldPct",
          "milestones",
          "monthlyContributionUsd",
          "principalUsd",
          "series",
          "years"
        ],
        "type": "object",
        "properties": {
          "principalUsd": {
            "type": "number",
            "format": "double"
          },
          "monthlyContributionUsd": {
            "type": "number",
            "format": "double"
          },
          "annualYieldPct": {
            "type": "number",
            "format": "double"
          },
          "years": {
            "type": "integer",
            "format": "int32"
          },
          "series": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectionPointResponse"
            }
          },
          "milestones": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MilestoneResponse"
            }
          }
        }
      },
...
    }
  }
}